Implemented solution for the issue #9.
Then loccur or loccur-no-highlight are called with universal prefix,
i.e. with C-u before the command, the currently selected value is
ignored.

Then people want this behavior by default, it is better wrap the call
to loccur with universal prefix, i.e. by implementing a helper function
like this:

(defun loccur-no-selection ()
  (interactive)
    (let ((current-prefix-arg 1))
        (call-interactively
             'loccur)))

And then just call this function instead of loccur.